Complete the styling for header section
Utilised flexbox to layout the navbar at the top of the header and
organise the hero text, sign up button and image below the navbar.

Added a couple of random images to images/ that are planned to be
incorporated into the website.

Added more font and color styling to the whole page in line with
colors-and-font-styles.png.

Minor modifications were made to the names of classes in index.html to
improve readability.